Moved ValidatePathError.ts to /core/src/lib/errors
