+++ packages/core/src/index.ts
@@ -28,7 +28,7 @@ import {TwaGenerator} from './lib/TwaGenerator';
 import {DigitalAssetLinks} from './lib/DigitalAssetLinks';
 import * as util from './lib/util';
 import {Result} from './lib/Result';
-import {ValidatePathError} from './lib/types/ValidatePathError';
+import {ValidatePathError} from './lib/errors/ValidatePathError';
 
 export {AndroidSdkTools,
   Config,
