+++ packages/core/src/lib/errors/ValidatePathError.ts
